Changed lazy to property operator
Changed the type of the lazy decorator from Method- to PropertyDecorator cause of the following error message when using lazy as documented: Unable to resolve signature of property decorator when called as an expression. (ts1240)
According to the implementation it makes more sense anyway since it only decorates a property.