Fix code example
I just found out about WatermelonDB an hour ago, so not sure if this is right. But shouldn't this be EnhancedComment instead of Comment?
Or is there an independent EnhancedPost after this. If that is the case, it would be better to have atleast one EnchancedComponent that is actually used, not just created.
Thanks